#164: use deleteReply to consolidate viewrepairs flow
