+++ source/rooms/workshop-tinker.js
@@ -47,7 +47,7 @@ module.exports = new RoomTemplate("Abandoned Forge",
 						.setEmoji(upgradeEmoji)
 						.setLabel(upgradeLabel)
 						.setDisabled(isUpgradeDisabled),
-					new ButtonBuilder().setCustomId("viewrepairs")
+					new ButtonBuilder().setCustomId("repair")
 						.setStyle(ButtonStyle.Primary)
 						.setEmoji(repairEmoji)
 						.setLabel(repairLabel)
